+++ src/testing/models.ts
@@ -1,3 +1,5 @@
+import { TracerEvent } from '../tracer';
+
 export interface Threshold {
   lt?: number;
   lte?: number;
@@ -8,7 +10,7 @@ export interface Threshold {
 export interface Evaluation {
   score: number;
   threshold?: Threshold;
-  metadata?: Record<string, string>;
+  metadata?: Record<string, unknown>;
 }
 
 export abstract class BaseTestEvaluator<TestCaseType, OutputType> {
@@ -19,5 +21,15 @@ export abstract class BaseTestEvaluator<TestCaseType, OutputType> {
   abstract evaluateTestCase(args: {
     testCase: TestCaseType;
     output: OutputType;
+  }): Evaluation | Promise<Evaluation>;
+}
+
+export abstract class BaseEventEvaluator {
+  abstract get id(): string;
+
+  maxConcurrency: number = 10;
+
+  abstract evaluateEvent(args: {
+    event: TracerEvent;
   }): Evaluation | Promise<Evaluation>;
 }
