+++ src/tracer/index.ts
@@ -1,7 +1,2 @@
 export { AutoblocksTracer } from './tracer';
-export {
-  BaseEventEvaluator,
-  Evaluation,
-  type TracerEvent,
-  type SendEventArgs,
-} from './models';
+export { Evaluation, type TracerEvent, type SendEventArgs } from './models';
