+++ src/tracer/models.ts
@@ -1,5 +1,6 @@
 import { Event } from '../client';
 import { Evaluation as TestingEvaluation } from '../testing';
+import { BaseEventEvaluator } from '../testing/models';
 import { ArbitraryProperties, PromptTracking } from '../types';
 
 export type TracerEvent = Omit<Event, 'id' | 'traceId'> & { traceId?: string };
@@ -13,16 +14,6 @@ export interface SendEventArgs {
   properties?: ArbitraryProperties;
   promptTracking?: PromptTracking;
   evaluators?: BaseEventEvaluator[];
-}
-
-export abstract class BaseEventEvaluator {
-  abstract get id(): string;
-
-  maxConcurrency: number = 10;
-
-  abstract evaluateEvent(args: {
-    event: TracerEvent;
-  }): Evaluation | Promise<Evaluation>;
 }
 
 export interface EvaluationWithIds extends Evaluation {
